feat(authorization): added authorization

+++ server/api/v1/summaries/index.delete.ts
@@ -2,6 +2,7 @@ import { useServerAuth } from '~~/server/utils/auth/useServerAuth'
 import { useServerSummaries } from '~~/server/utils/core/useServerSummaries'
 import {z} from 'zod'
 import { useServerFiles } from '~~/server/utils/core/useServerFiles'
+import { summaryAbilities } from '#shared/abilities/summaryAbilities'
 
 const QuerySchema = z.object({
     ids: z.union([z.uuid(), z.array(z.uuid())])
@@ -19,6 +20,10 @@ export default defineEventHandler(async (event) => {
     const $file = useServerFiles()
 
     const { ids } = await getValidatedQuery(event, QuerySchema.parse)
+
+    const s: Summary[] = await $sum.getSummariesById(ids)
+
+    await authorize(event, summaryAbilities().readOrUpdateOrDeleteSummary, s)
 
     await $file.deleteFilesAndBlobsFromSummaries(user.user.id, ids)
